+++ NEWS.md
@@ -1,6 +1,11 @@
 Pentobi Release Notes
 =====================
 
+Version 28.0-dev (main)
+-----------------------
+
+* 
+
 Version 27.1-dev (v27-fixes)
 ----------------------------